requirements.txt
langchain==0.1.0
langgraph==0.0.40
langchain-community==0.0.20
langchain-experimental==0.0.50
ollama==0.1.7
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4==4.12.2
duckduckgo-search==3.9.6
schedule==1.2.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1